debug the parser (syntax analyzer)

+++ src/common.h
@@ -21,12 +21,12 @@
 typedef char                   bool;
 typedef char                   byte;
 typedef char                   int8;
-typedef short int              int16;
-typedef long int               int32;
+typedef short                  int16;
+typedef int                    int32;
 typedef long long int          int64;
 typedef unsigned char          uint8;
-typedef unsigned short int     uint16;
-typedef unsigned long int      uint32;
+typedef unsigned short         uint16;
+typedef unsigned int           uint32;
 typedef unsigned long long int uint64;
 typedef float                  float32;
 typedef double                 float64;
